# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Betting Intelligence API",
    "description": "AI-powered basketball betting recommendations. Real odds from TheOddsAPI + ESPN scraper, ML predictions, Kelly staking, and subscription management.\n\n## Data Sources\n- **TheOddsAPI** (tier 1): Primary source, requires API key\n- **ESPN Public API** (tier 2): Chrome-based stealth scraper, no key needed — automatic fallback when TheOddsAPI is unavailable",
    "version": "0.2.0",
    "contact": {
      "name": "Betting Intelligence",
      "url": "http://localhost:8000"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/bets": {
      "get": {
        "summary": "List Bets with Filtering",
        "description": "Returns all bets with optional filtering by league, bet type, minimum edge, and result limit.",
        "operationId": "apiBets",
        "tags": ["Data"],
        "parameters": [
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "all" },
            "description": "Filter by league (e.g., NBA, WNBA) or 'all'"
          },
          {
            "name": "bet_type",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "all" },
            "description": "Filter by bet type or 'all'"
          },
          {
            "name": "min_edge",
            "in": "query",
            "required": false,
            "schema": { "type": "number", "default": 0.0 },
            "description": "Minimum edge percentage threshold"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 },
            "description": "Maximum number of results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of bet objects sorted by edge descending",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Bet" }
                }
              }
            }
          }
        }
      }
    },
...
    "/all-bets": {
      "get": {
        "summary": "All Bets Page",
        "description": "All bets with HTMX sorting/filtering by league, bet type, minimum edge, and sort order.",
        "tags": ["Pages"],
        "parameters": [
          { "name": "league", "in": "query", "schema": { "type": "string", "default": "all" } },
          { "name": "bet_type", "in": "query", "schema": { "type": "string", "default": "all" } },
          { "name": "min_edge", "in": "query", "schema": { "type": "number", "default": 0.0 } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["edge", "stake", "confidence"], "default": "edge" } }
        ],
        "responses": {
          "200": {
            "description": "HTML all bets page"
          }
        }
      }
    },
    "/clear-picks": {
      "get": {
        "summary": "Clear Picks Page",
        "description": "High-confidence clear picks with HTMX filtering by league, bet type, and minimum edge.",
        "tags": ["Pages"],
        "parameters": [
          { "name": "league", "in": "query", "schema": { "type": "string", "default": "all" } },
          { "name": "bet_type", "in": "query", "schema": { "type": "string", "default": "all" } },
          { "name": "min_edge", "in": "query", "schema": { "type": "number", "default": 0.0 } }
        ],
        "responses": {
          "200": {
            "description": "HTML clear picks page"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Bet": {
        "type": "object",
        "properties": {
          "game_id": { "type": "string" },
          "game_date": { "type": "string" },
          "matchup": { "type": "string" },
          "league": { "type": "string" },
          "bet_type": { "type": "string" },
          "bet_type_display": { "type": "string" },
          "bet_type_icon": { "type": "string" },
          "bet_side": { "type": "string" },
          "edge_pct": { "type": "number" },
          "stake_dollars": { "type": "number" },
          "kelly_fraction": { "type": "number" },
          "win_probability": { "type": "number" },
          "confidence": { "type": "string" },
          "is_clear_pick": { "type": "boolean" },
          "reasoning": { "type": "string" },
          "model_name": { "type": "string" },
          "market_line": { "type": "number", "nullable": true },
          "predicted_value": { "type": "number" }
        }
      },
...
    }
  },
...
}